support jaeger on openshift for algorithm-queue
